refactor(test): 테스트 데이터 분리
